requirements.txt
setuptools
Django>=5.0.4
Pillow==10.3.0
django-polymorphic==3.1.0
django-import-export==3.3.9
requests>=2.31.0
ipython==8.18.1
lxml==5.1.0
six==1.16.0
typing-extensions==4.9.0
urllib3>=1.26.18
gunicorn==22.0.0
dj-database-url
psycopg2==2.9.9
sentry-sdk[django]
python-dotenv